SetUp Chakra Ui

+++ Backend/controllers/products.js
@@ -77,6 +77,13 @@
  export const deleteProduct = async (req, res) => {
      try {
          const {id} = req.params;
+
+         if(!mongoose.Types.ObjectId.isValid(id)){
+            return res.status(404).json({
+                success: false,
+                message: 'Product not found'
+            });
+         }
  
          const product = await Product.findByIdAndDelete(id);
           return res.status(200).json({
